backon
Trait Sleeper

pub trait Sleeper: 'static {
    type Sleep: Future<Output = ()>;

    // Required method
    fn sleep(&self, dur: Duration) -> Self::Sleep;
}

A sleeper is used to generate a future that completes after a specified duration.

Required Associated Types

type Sleep: Future<Output = ()>

The future returned by the sleep method.

Required Methods

fn sleep(&self, dur: Duration) -> Self::Sleep

Create a future that completes after a set period.